Add files via upload
A set of functions for Ridge regression. There are also functions for end of sample cross-validation and a function for creating a vector of values for the shrinkage parameter lambda (or v). After generating a set of lambda values, the optimal value can be chosen through end of sample cv. There are two function that apply recursive one-step ahead forecast to a dataset. One of those function does end of sample CV at each point in time, the other one only uses on lambda value for the whole forecasting exercise.